+++ server/README.md
@@ -1,0 +1,3 @@
+# Server
+
+The Flask server, works as the relay station between Student-end (GUI) and Teacher-end (Monitor & Database).
