+++ teacher/README.md
@@ -1,3 +1,7 @@
 # Teacher Monitor
 
-The monitor view on the Teacher-end.
+The monitor view and controller of the Teacher-end.
+
+## Database
+
+The database of statuses and grades of students are also stored in the Teacher-end.
